// src/core/applications/user/dto/update-user.dto.ts
export interface UpdateUserDto {
  fullName?: string;
  phone?: string;
  nickname?: string;
  city?: string;
  gender?: string;
  ageGroup?: string;
  nationality?: string;
  employmentSector?: string;
  profession?: string;
  personalEmail?: string;
  loyaltyTier?: string;
  subscriptionTier?: string;
  profileStatus?: string;
  roles?: string[];
}